// apps/web/src/http/api.schemas.ts
export type GetProjects200ProjectsItemOwner = {
  id: string
  /** @nullable */
  name: string | null
  /** @nullable */
  avatarUrl: string | null
}

export type GetProjects200ProjectsItem = {
  id: string
  description: string
  name: string
  slug: string
  /** @nullable */
  avatarUrl: string | null
  organizationId: string
  ownerId: string
  createdAt: string
  owner: GetProjects200ProjectsItemOwner
}